backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
aiofiles==23.2.1

# ML & Inference
torch>=2.0.0
torchvision>=0.15.0
onnxruntime>=1.16.0
# onnxruntime-gpu>=1.16.0  # Uncomment for GPU

# Image/Video Processing
opencv-python-headless>=4.8.0
pillow>=10.0.0
numpy>=1.24.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1

# Utilities
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Advanced Analytics (HuggingFace)
huggingface_hub>=0.20.0
gradio_client>=0.10.0

# PDF Report Generation
reportlab>=4.0.0
